requirements.txt
# === Core Data Science ===
numpy>=1.24.3
pandas>=2.0.0
matplotlib>=3.7.1
seaborn>=0.12.2
scipy>=1.10.1

# === Image Processing ===
opencv-python>=4.7.0
Pillow>=9.5.0

# === Utilities ===
tqdm>=4.65.0
psutil>=5.9.5
absl-py>=1.4.0

# === Deep Learning (TensorFlow) ===
# 用于 DeepLab 语义分割
tensorflow>=2.13.0

# === Deep Learning (PyTorch) ===
# 用于 SAM2 与 CLIP
torch>=2.0.0
torchvision>=0.15.0

# === Vision & Language Models ===
# 用于 CLIP 模型加载与推理
transformers>=4.30.0

# === Geospatial & Geometry ===
# 用于 Mask 轮廓转 WKT 多边形
shapely>=2.0.0

# === Excel Support (Pandas) ===
# 用于导出最终分析结果表
openpyxl>=3.1.2

# ==========================================
# 注意：SAM2 依赖不在此文件中。
# 请进入 sam2_repo 目录单独执行 pip install -e .
# ==========================================